move pages to their own packges while keeping app starters in the top level of gui
